# py/gc: Speed up incremental GC cycles by tracking the high watermark.