@damz 2022-12-15
Some performance measurements. Using this trivial script:

```python
import gc
import time

for i in range(30):
    start = time.ticks_ms(); gc.collect(); print(time.ticks_diff(time.ticks_ms(), start))
```

On the ESP32-S3 with 2 MB of SPIRAM, right after boot:

* Before: ~20ms
* After: <1ms

On the Unix port with a 1 GB heap:

* Before: ~47ms
* After: <1ms

In our real-life application (the [Ribbit Network frog sensor](https://github.com/Ribbit-Network/ribbit-network-frog-software) on the same ESP32-S3 with 2 MB of SPIRAM), we run GC every minute and use ~43 kB of memory after the GC completes, this cuts the GC duration as well:

* Before: 41-42ms
* After: 22-23ms