@jimmo 2022-12-16
This is +56 bytes on PYBV11. Performance tests look positive too. Note that the performance benchmarks don't explicitly test the GC, but many tests will trigger collections.

```
$ ./run-perfbench.py -s ~/gc-hwm-pybv11-baseline ~/gc-hwm-pybv11-damz 
diff of scores (higher is better)
N=100 M=100                /home/jimmo/gc-hwm-pybv11-baseline -> /home/jimmo/gc-hwm-pybv11-damz         diff      diff% (error%)
bm_chaos.py                    345.65 ->     352.87 :      +7.22 =  +2.089% (+/-0.00%)
bm_fannkuch.py                  75.36 ->      75.70 :      +0.34 =  +0.451% (+/-0.00%)
bm_fft.py                     2291.07 ->    2333.58 :     +42.51 =  +1.855% (+/-0.00%)
bm_float.py                   5672.50 ->    5710.00 :     +37.50 =  +0.661% (+/-0.01%)
bm_hexiom.py                    46.20 ->      46.46 :      +0.26 =  +0.563% (+/-0.00%)
bm_nqueens.py                 4218.20 ->    4246.61 :     +28.41 =  +0.674% (+/-0.00%)
bm_pidigits.py                 627.03 ->     620.55 :      -6.48 =  -1.033% (+/-0.13%)
misc_aes.py                    406.43 ->     409.94 :      +3.51 =  +0.864% (+/-0.01%)
misc_mandel.py                3035.28 ->    3125.27 :     +89.99 =  +2.965% (+/-0.01%)
misc_pystone.py               2324.40 ->    2349.33 :     +24.93 =  +1.073% (+/-0.01%)
misc_raytrace.py               362.17 ->     367.64 :      +5.47 =  +1.510% (+/-0.00%)
```