@jimmo 2022-12-16
As a very rudimentary attempt to make the benchmarks more sensitive to GC performance, I added a `gc.threshold(30000)` (which is ~1/3 of the available heap on PYBV11).

First, this diff shows the effect that has on the tests (without the diff from this PR).

```
$ ./run-perfbench.py -s ~/gc-hwm-pybv11-baseline ~/gc-hwm-pybv11-threshold30-baseline 
diff of scores (higher is better)
N=100 M=100                /home/jimmo/gc-hwm-pybv11-baseline -> /home/jimmo/gc-hwm-pybv11-threshold-baseline         diff      diff% (error%)
bm_chaos.py                    345.65 ->     324.05 :     -21.60 =  -6.249% (+/-0.00%)
bm_fannkuch.py                  75.36 ->      71.60 :      -3.76 =  -4.989% (+/-0.00%)
bm_fft.py                     2291.09 ->    2149.10 :    -141.99 =  -6.197% (+/-0.01%)
bm_float.py                   5672.70 ->    4892.95 :    -779.75 = -13.746% (+/-0.00%)
bm_hexiom.py                    46.20 ->      45.51 :      -0.69 =  -1.494% (+/-0.00%)
bm_nqueens.py                 4218.15 ->    3995.01 :    -223.14 =  -5.290% (+/-0.00%)
bm_pidigits.py                 628.75 ->     431.93 :    -196.82 = -31.303% (+/-0.30%)
core_import_mpy_multi.py       513.34 ->     497.46 :     -15.88 =  -3.093% (+/-0.01%)
core_import_mpy_single.py       64.71 ->      60.42 :      -4.29 =  -6.630% (+/-0.01%)
core_qstr.py                    64.39 ->      64.37 :      -0.02 =  -0.031% (+/-0.00%)
core_yield_from.py             353.02 ->     353.02 :      +0.00 =  +0.000% (+/-0.00%)
misc_aes.py                    406.43 ->     406.45 :      +0.02 =  +0.005% (+/-0.01%)
misc_mandel.py                3035.28 ->    2903.97 :    -131.31 =  -4.326% (+/-0.00%)
misc_pystone.py               2324.45 ->    2239.00 :     -85.45 =  -3.676% (+/-0.01%)
misc_raytrace.py               362.16 ->     339.32 :     -22.84 =  -6.307% (+/-0.00%)
```

Now with this PR added:

```
$ ./run-perfbench.py -s ~/gc-hwm-pybv11-threshold30-baseline ~/gc-hwm-pybv11-threshold30-damz 
diff of scores (higher is better)
N=100 M=100                /home/jimmo/gc-hwm-pybv11-threshold-baseline -> /home/jimmo/gc-hwm-pybv11-threshold-damz         diff      diff% (error%)
bm_chaos.py                    324.05 ->     340.00 :     +15.95 =  +4.922% (+/-0.01%)
bm_fannkuch.py                  71.60 ->      74.22 :      +2.62 =  +3.659% (+/-0.00%)
bm_fft.py                     2149.10 ->    2207.07 :     +57.97 =  +2.697% (+/-0.01%)
bm_float.py                   4892.95 ->    5026.46 :    +133.51 =  +2.729% (+/-0.01%)
bm_hexiom.py                    45.51 ->      45.83 :      +0.32 =  +0.703% (+/-0.00%)
bm_nqueens.py                 3995.01 ->    4164.99 :    +169.98 =  +4.255% (+/-0.00%)
bm_pidigits.py                 431.93 ->     455.80 :     +23.87 =  +5.526% (+/-0.59%)
core_import_mpy_multi.py       497.46 ->     507.83 :     +10.37 =  +2.085% (+/-0.01%)
core_import_mpy_single.py       60.42 ->      62.65 :      +2.23 =  +3.691% (+/-0.00%)
core_qstr.py                    64.37 ->      64.30 :      -0.07 =  -0.109% (+/-0.00%)
core_yield_from.py             353.02 ->     354.93 :      +1.91 =  +0.541% (+/-0.01%)
misc_aes.py                    406.45 ->     411.28 :      +4.83 =  +1.188% (+/-0.01%)
misc_mandel.py                2903.97 ->    3083.66 :    +179.69 =  +6.188% (+/-0.00%)
misc_pystone.py               2239.00 ->    2271.89 :     +32.89 =  +1.469% (+/-0.02%)
misc_raytrace.py               339.32 ->     354.83 :     +15.51 =  +4.571% (+/-0.01%)
```