@jkhax0r 2023-03-26
I'm on an ESP32S3 WROOM module with 8MB octal SPIRAM on custom hardware.  Found this PR b/c was struggling with ~105ms gc.collect() times when I'm only actually using about 80KB of RAM (after GC)

Running gc.collect() every 250ms I was seeing ~100ms. And then ~105ms when I run every e.g. 1-2 minutes.

I manually merged this PR and went from that ~100ms down to ~46ms when I run gc.collect() from a 250ms loop.  And the 105ms down to 50ms when run every 1-2 minutes.  

Thanks for the PR... so simple for such a huge improvement!  

Do these numbers seem right to you guys?   

Prior to this PR I was seeing ~60ms IIRC with just the below loop from the REPL with no code running.  Now at 0 to 1ms.  Nice.
```
for i in range(0,100):
    t = utime.ticks_ms()
    gc.collect()
    print('t=', utime.ticks_ms() - t)
```