@peterhinch 2023-03-26
For comparison an RP2 Pico with standard firmware shows 2-3ms. On past experience a Pyboard would show 1-2ms.

Incidentally, while OK in quick tests, in production code subtracting `ticks` values is unsafe because of rollover: use `ticks_diff()`.